Add new feature: blog posts
In order to render blog posts, I had to install the package
next-mdx-remote. With this package, I was able to render the
contents of MDX docs. I also had to edit the Tailwind CSS config
file to allow me to use Tailwind CSS classes for HTML elements
within MDX docs.