Update Next.js config and add Docker files
I updated the Next.js config by setting the output to standalone.
This allows Next.js to create a standalone folder that copies only
the files needed in a production environment.

I've also included a Dockerfile since I will be creating an image
to run my app in a docker container.